rename lb test to health
